requirements.txt
python-dotenv
pydantic
mem0ai
httpx
fastapi
uvicorn
